requirements.txt
asgiref==3.2.7
Django==3.0.5
django-widget-tweaks==1.4.8
Pillow==7.1.1
pytz==2019.3
sqlparse==0.3.1
asgiref==3.2.7
Django==3.0.5
django-appconf==1.0.4
django-image-cropping==1.4.0
django-widget-tweaks==1.4.8
easy-thumbnails==2.7
Pillow==7.1.1
pytz==2019.3
sqlparse==0.3.1
amqp==2.5.2
asgiref==3.2.7
billiard==3.6.3.0
celery==4.4.2
Django==3.0.5
django-appconf==1.0.4
django-celery-email==3.0.0
django-image-cropping==1.4.0
django-widget-tweaks==1.4.8
easy-thumbnails==2.7
importlib-metadata==1.6.0
kombu==4.6.8
Pillow==7.1.1
psycopg2==2.8.5
python-http-client==3.2.7
pytz==2019.3
redis==3.4.1
sendgrid==6.2.2
sqlparse==0.3.1
vine==1.3.0
zipp==3.1.0
amqp==2.5.2
asgiref==3.2.7
billiard==3.6.3.0
celery==4.4.2
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
cryptography==2.9
defusedxml==0.6.0
Django==3.0.5
django-appconf==1.0.4
django-celery-email==3.0.0
django-image-cropping==1.4.0
django-widget-tweaks==1.4.8
easy-thumbnails==2.7
idna==2.9
importlib-metadata==1.6.0
kombu==4.6.8
oauthlib==3.1.0
Pillow==7.1.1
psycopg2==2.8.5
pycparser==2.20
PyJWT==1.7.1
python-http-client==3.2.7
python-social-auth==0.3.6
python3-openid==3.1.0
pytz==2019.3
redis==3.4.1
requests==2.23.0
requests-oauthlib==1.3.0
sendgrid==6.2.2
six==1.14.0
social-auth-app-django==3.1.0
social-auth-core==3.3.3
sqlparse==0.3.1
urllib3==1.25.9
vine==1.3.0
zipp==3.1.0
